Rename: Autocomplete.vue > SearchBar.vue / UPDATE: Home.vue / Remove: HelloWorld.vue
